+++ cmd/wyze-bridge/main.go
@@ -28,7 +28,7 @@ import (
 )
 
 // Version is set at build time via ldflags.
-var Version = "4.6.4"
+var Version = "4.6.5"
 
 func main() {
 	// Load configuration